FELIX-6212 : Handle bad resolve context for substitute packages
The fix for FELIX-6212 made it such that resolve contexts that do not
represent substituted exports accurately would now totally fail with a
runtime exception.  Instead of totally failing this fix allows them to
continue and ignores the blame when no sources can be found for the
package from the provider.