[FELIX-6274] - Stop supporting pre R7 LogService for the Logger
The OSGi R7 LogService introduced the LoggerFactory which allows SCR to
get the Logger for another bundle without having to resort to using the
extendee bundles context to get the LogService.  This allows SCR to use
the log package SCR is wired to without having to worry about the
package that the extendee is wired to for log.  It also simplifies much
with how SCR logs and it keeps the log entries for all bundles that SCR
extends flowing through the same LogService that SCR uses instead of
disjointly though all the different LogServices the extendees may be
wired to.

Change-Id: I4f1169adf089b34fb460f4b2f40607c17948c909
Signed-off-by: Thomas Watson <tjwatson@us.ibm.com>