[configurator] Unchanged bundle does not obey force policy
The force policy indicates that a bundle's configuration should
always override any local changes. 

However, the code contained an optimization that if the configuration
in a bundle was unchanged, it would ignore its configuration
completely, ignoring any force policies.

This patch overrides the optimization in ConfigList for records
where the policy is set to force.

- A test is added that failed on the current bundle and now passes
- The OSGi Test Suite has been run on the code, the XML file is
  included in the patch