FELIX-6317 - Ensure the order of dynamic binding follows service rank
If multiple threads register higher ranked services than
the currently bound service then a single thread is
selected to do the work of binding the highest ranked
service. The other threads just queue up the possible
service to be bound by the winning thread.

This ensures that the component will not
get out of order bindings for the various service
registration when a component has a dynamic, greedy reference
with cardinality of 0..1 or 1..1

A new testcase is added that registers 100 services with
different ranks using 10 threads. The test then confirms
that the highest ranked service is the current one bound
to the component. Theoretically the test could pass before
this fix, but multiple runs without the fix never resulted
in a passing test.