FELIX-6321 - keep existing LogManager on config update
Replacing the existing LogManager with a new one on
configuration updates causes the existing loggers for
each component to stop working properly.  This is a quick
fix to avoid creating a new LogManager on config updates
of SCR.  If dynamic updates is needed for the ds.log.extension
property then more work is needed.  For now this is a
one time setting only configured through context properties