Introduced thread-safety while updating field during deactivation phase
The problem occurred when multiple injected services tried to be removed from the component's injection. It is highly likely because of the usage of TreeMap in such concurrent environment. Since TreeMap has never been written to be used in a concurrent environment, we should use ConcurrentSkipListMap instead as all the features of TreeMap have been introduced in ConcurrentSkipListMap and it is very much suitable for concurrent environment.
Introduced thread-safety while updating field during deactivation phase