FELIX-6581 - use system context to clean up logger domains
Always using the system context to clean up the logger domians.
It is more simple to always use the system context for this
purpose instead of conditionally use it according to the
ds.global.extender configuration.  There is no harm to using
always using the system context here because we only use it to
listen for stopped bundles to clean up used domains.