FELIX-6616 : reduce timing window for binding dynamic greedy refs
For 1..1 and 0..1 dynamic greedy references it is possible that
the activating thread will bind a lower ranked service while
another thead is binding a higher ranked service.  This can
result in some strange orders for bind methods.

This change reduces the timing window such that the activating
thread will immediately bind the higher ranked service as long
as another thread has indicated that it will bind the higher
ranked service. This still results in some extra bind calls
for the same higher ranked service but greatly reduces the
window where the activating thread will bind a lower ranked
service while the tracking thread is trying to bind a higher
ranked service.

This change moves the check for what ref to bind from the
activating thread to just before it will call bind on the
component object being activated. There is still a small
window where another thread may register a higher ranked
service right after the activating thread gets the ref
to bind and when it actually calls bind. If this proves
to be problematic then we may need an additional check
after bind during activate that makes sure the latest
tracked service is what actually got bound during activate.