Updated to use HTTP optionally in HealthCheck (#192)
* Imported OSGi HTTP Package Optionally

fixes FELIX-6565

* Replaced use of HTTP Service with HTTP Whiteboard

There was no runtime issue to use the old pattern of using `HttpService` to register the servlets, but dependency to `HttpService` introduces a mandatory requirement to `osgi.service` namespace. 

`osgi.service;filter:="(objectClass=org.osgi.service.http.HttpService)";effective:=active`

This requirement doesn't play any role in the runtime but it plays a significant role in composing systems. If we use bndtools and use this bundle as an initial requirement to resolve the dependencies, this will fail as the `requirement` needs to be satisfied first. We also use similar mechanism in our company to compose systems using `OSGi Resolver` which would not work if this bundle comprises such mandatory dependency to `HttpService`. That's why, a cleaner approach would be to go for **OSGi HTTP Whiteboard** where dependency to `HttpService` is not required at all, and it now optionally requires HTTP whiteboard.