FELIX-6633 : Dynamic references may miss events while causing deactivation
Ensure no new add events got missed when deactivating on removed service

This change follows the same pattern as other reference types but
also does it for dynamic references when they become unsatisifed.

The fix is to simply cause the dependency manager to be checked
again after the removed event gets processed to ensure the dependency
is not now satisified.