FELIX-6744 Configure policy for maven release plugin - Configure OddEvenVersionPolicy based on the undocumented rule
The reason is the difference in how maven and osgi orders versions with a qualifier. In maven 1.0.19-SNAPSHOT is lower than 1.0.19 while in OSGi it is higher. That sometimes confuses tooling.
For example, in case you have already installed the snapshot version and want to install the final version and use OSGi tooling based on "install the highest version" it will keep the snapshot version installed.
To avoid all of that we decided to just use even numbers for releases.
Its cheap and avoids all these problems.