[http] extract Multipart code into new ServletRequestMultipartWrapper
Mostly in the usage of felix-jetty-light there is the cases where you do
not need the multipart support and can remove the dependencies of
common-fileupload and common-io. To be able to do this we could not load
a instantiate a class that imports common-fileupload classes. So we have
to free the old class ServletRequestWrapper of this imports and add a
ServletRequestMultipartWrapper that still have the multipart support.

Signed-off-by: Stefan Bischof <stbischof@bipolis.org>