Do not print an error when a wildcard is used in component header
Currently SCR prints an error if one uses a wildcard in the header but
nothing matches. But because fragemnts are not allowed to provide an own
header, only the host can enable this.

So this now adds an additional check before logging an error that
prevents an error from being printed if a wildcard is provided and
prints a trace message instead.