[FELIX-6814] Improve Reproducible Build Support
Use the value of Maven's project.build.outputTimestamp property as the value for the -reproducible instruction.

Timestamps, formatted in ISO-8601 or in number of seconds since the epoch, are both supported by bnd tools. The value is also already sanitized by MavenArchiver::parseOutputTimestamp which also allows only numeric timestamps or timestamps in ISO-8601 format.